+++ src/app.js
@@ -1,3 +1,19 @@
+import {Modal} from "./modal";
+import {Skill} from "./skill";
+
+const lt = /</g,
+    gt = />/g,
+    ap = /'/g,
+    ic = /"/g
+
+export function checkXss(str) {
+    return str.toString()
+        .replace(lt, "&lt;")
+        .replace(gt, "&gt;")
+        .replace(ap, "&#39;")
+        .replace(ic, "&#34;")
+}
+
 // document.addEventListener("DOMContentLoaded", function (event) {
 //     document.body.addEventListener("click", function () {
 //         if (document.body.classList.contains("light-theme")) {
@@ -7,6 +23,7 @@
 //         }
 //     })
 // });
+
 document.addEventListener("DOMContentLoaded", function (event) {
     const skill = new Skill();
     const into = document.querySelector(".skills_wrapper");
@@ -17,6 +34,7 @@ document.addEventListener("DOMContentLoaded", function (event) {
     Modal.initialize();
 
     const form = document.getElementById("form");
+
     function retrieveFormValue(event) {
         event.preventDefault();
 
@@ -28,56 +46,3 @@ document.addEventListener("DOMContentLoaded", function (event) {
 
     form.addEventListener('submit', retrieveFormValue);
 });
-
-class Modal {
-    static initialize() {
-        const modal = document.querySelector(".modal");
-
-        document.getElementById("skill-btn").addEventListener("click", function () {
-            modal.classList.add('active');
-        })
-        document.querySelector(".modal_close-button").addEventListener("click", function () {
-            modal.classList.remove('active');
-        })
-    }
-}
-
-class Skill {
-    lt = /</g
-    gt = />/g
-    ap = /'/g
-    ic = /"/g
-
-    insert(element, name, percents) {
-        name = name.toString()
-            .replace(this.lt, "&lt;")
-            .replace(this.gt, "&gt;")
-            .replace(this.ap, "&#39;")
-            .replace(this.ic, "&#34;")
-        const namePercentsButton = `<div class="skill_description"><div class="skills_info">${name}</div><div class="skills_info">${percents}%</div></div>`
-        const progressBar = `<div class="progress"><div class="progress-value" style="max-width: ${percents}%"></div></div>`
-        const wrapper = `<div class="skills_node"><div>${namePercentsButton}</div><div>${progressBar}</div></div>`
-
-        const btn = document.createElement('input');
-        btn.type = "image"
-        btn.src = 'assets/minus.png';
-        btn.width = 30;
-        btn.height = 30;
-        btn.style.marginTop = "25px";
-        btn.onclick = function () {
-            remove(btn);
-        }
-
-        const skillsButton = document.createElement("div");
-        skillsButton.classList.add("skills_button");
-        skillsButton.innerHTML = `${wrapper}`;
-        skillsButton.appendChild(btn)
-
-        element.appendChild(skillsButton);
-    }
-}
-
-function remove(e) {
-    const parent = document.querySelector(".skills_wrapper");
-    parent.removeChild(e.parentNode);
-}
